changed framed Animation to use float for all times, builders keep the same but transform int time in milliseconds to float time in seconds

+++ commons-gdx-core/src/main/java/com/gemserk/animation4j/FrameAnimation.java
@@ -7,6 +7,6 @@ public interface FrameAnimation {
 
 	int getCurrentFrame();
 
-	void update(int delta);
+	void update(float delta);
 
 }
